requirements.txt
selenium
openpyxl
PyYAML
PyMySQL
loguru
pytest
Faker
allure-pytest
jsonpath